Zmiana nazw istniejacych testow, dodanie kolejnych

+++ ServerApplication/src/libs/dao/CChecksumList.h
@@ -6,7 +6,7 @@
 /**
  * @brief The CChecksumList class represents list of file checksum.
  *
- * @details @todo dopisać
+ * @todo dopisać
  */
 class CChecksumList {
   public:
@@ -16,31 +16,38 @@ class CChecksumList {
     CChecksumList();
     /**
      * @brief AddFileChecksum add new file data checksum
-     * to checksum list.
+			 *				to checksum list.
      *
      * @param aChecksum is new checksum value.
      */
-    void AddFileChecksum(int aChecksum);
+		void AddFileChecksum(int16_t aChecksum);
 
     /**
      * @brief PopulateChecksumFromDB function add checksum file list
-     * of actually server files.
+			 *				of actually server files.
      */
     void ReceiveChecksumsFromDB();
 
     /**
-     * @brief CheckFileCheckSum compare checksum in parameter to checksums in server
-     * which we import to list in class CCheckSumList.
+			 * @brief CheckFileCheckSum compare checksum in parameter to
+			 *				checksums in server which we import to list in
+			 *				class CCheckSumList.
      *
      * @param aCheckSum is checksum we compare to
-     * checksum from server.
+			 *				checksum from server.
      *
-     * @return TRUE if checksum is the same in another in server.
+			 * @return True if checksum is the same in another in server.
      */
-    bool CheckFileChecksum(int aChecksum);
+		bool CheckFileChecksum(int16_t aChecksum);
+
+		/**
+		 * @brief UpdateChecksumList update list of checksums.
+		 */
+		//void UpdateChecksumList();
 
     /**
-     * @brief Clear function is a facade for clear function in Qlist class.
+			 * @brief Clear function is a facade for clear function in
+			 *				Qlist class.
      */
     void Clear();
